// src/api/generated/model/kindEnum.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

/**
 * * `kit` - Kit
 * * `component` - Component
 */
export type KindEnum = typeof KindEnum[keyof typeof KindEnum];


export const KindEnum = {
  kit: 'kit',
  component: 'component',
} as const;